requirements.txt
python-dotenv==1.0.1
mysql-connector-python==9.2.0